requirements.txt
smmap==5.0.2
sniffio==1.3.1
streamlit==1.45.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
streamlit
python-dotenv
autogen
openai
langchain
chromadb
langchain_community
